feat: Add axios as a dependency
This commit adds the "axios" package as a new dependency in the "package.json" file. The version added is "^1.6.1". Axios is used for making HTTP requests and will be utilized in the codebase.

The changes include modifying the "dependencies" section of the "package.json" file to include "@actions/github": "^6.0.0", and adding a new line with "axios": "^1.6.1".

Additionally, in the "downloadFile.ts" file, there are changes made to import statements and code logic related to downloading files using axios instead of https module.

Other changes include updates to yarn.lock file with new versions of dependencies like form-data, combined-stream, mime-db, mime-types, follow-redirects, asynckit, delayed-stream.

These changes were made to enhance functionality and improve performance in handling HTTP requests within the codebase.