feat: Add setOutput function for setting GitHub outputs
This commit adds a new function called `setOutput` to the `downloadFileRunner.ts` file. This function is used to set GitHub outputs and is a temporary hack until the core actions library catches up with GitHub's new recommendations. It appends the key-value pair to the `GITHUB_OUTPUT` environment variable, which is expected to be set. If the variable is not set, an error will be thrown.

The purpose of this change is to provide a workaround for setting GitHub outputs in cases where the core actions library does not support it yet.